refactor(grammars): Extract shared Python scanner state to separate file
- Move pyDelimiter, pythonScannerState, and serialization logic to python_scanner_shared.go
- Update Python, Mojo, and Starlark scanners to call shared functions directly
- Narrow python_scanner.go build tags to Python-only to stop metadata leakage in derivative builds
- Add regression test to block Python spec registration during derivative-only subset builds
- Wire subset_build_sweep.sh into CI to validate tag coverage automatically

Buckley-Change-Hash: sha256:66f62136dfd2f1d8f2b01fc9be28205af28c0b25e836326279da4d72a581ee73
Buckley-Change-Stats: files=7 insertions=179 deletions=124 binaries=0
